+++ settings.py
@@ -57,6 +57,11 @@ MIDDLEWARE = (
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 )
 
+USE_TZ = True
 TIME_ZONE = "UTC"
 
 LOCALE_PATHS = [str(BASE_PATH) + "/django_pony_express/locale"]
+
+
+# Pony express
+DJANGO_PONY_EXPRESS_LOG_RECIPIENTS = False