+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changelog
 
+* *2.1.0* (2024-05-27)
+  * Added `ThreadEmailService` for simple async sending of emails
+  * Added basic logging with privacy configuration to mail class
+  * Restructured documentation
+  * Restructured unit-tests
+  * Minor test improvements
+
 * *2.0.0* (2024-04-11)
   * Dropped Django 3.2 & 4.1 support (via `ambient-package-update`)
   * Internal updates via `ambient-package-update`